validator/requirements.txt
pymongo~=4.1.1
python-decouple==3.3